tests/wpt/tests/webrtc-encoded-transform/sframe-transform-readable.html

<!doctype html>
<html>
    <head>
        <meta charset="utf-8">
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
    </head>
    <body>
        <iframe src="." id="frame"></iframe>
        <script>
promise_test(async (test) => {
    const frameDOMException = frame.contentWindow.DOMException;
    const transform = new frame.contentWindow.SFrameTransform;
    frame.remove();
    assert_throws_dom("InvalidStateError", frameDOMException, () => transform.readable);
});
       </script>
    </body>
</html>
